Rename actions folders to match the API path
