+++ SingularityUI/app/actions/api/auth.es6
@@ -1,0 +1,6 @@
+import { buildApiAction } from './base';
+
+export const FetchUser = buildApiAction(
+  'FETCH_USER',
+  {url: '/auth/user'}
+);
